NEW - Graphical trainer sprite selection window
